+++ app/src/main/java/com/jaf/recipebook/db/recipes/RecipeDao.java
@@ -4,6 +4,7 @@ import androidx.room.Dao;
 import androidx.room.Delete;
 import androidx.room.Insert;
 import androidx.room.Query;
+import androidx.room.Relation;
 import androidx.room.Update;
 
 import java.util.List;
@@ -22,7 +23,7 @@ public interface RecipeDao {
     int deleteRecipe(RecipesModel model);
 
     @Query("SELECT * FROM recipes WHERE id = :id")
-    Single<RecipesModel> getRecipe(int id);
+    Single<RecipesModel> getRecipe(long id);
 
     @Query("SELECT * FROM recipes ORDER BY name ASC")
     List<RecipesModel> getAllRecipes();
